requirements.txt
python-telegram-bot>=21.0
Flask>=3.0
python-dotenv>=1.0

